refactor(weatherReport): extract WeatherReport shape into its own module
Candidate 2 from architecture review: normalizeToOwmShape built an
OWM-shaped object with no type, no module owning it. Every consumer read
it by key path (data.current.main.temp, data.forecast.list[0].dt). The
shape leaked across every seam.

Extracted normalizeToOwmShape into src/weatherReport.js — the canonical
WeatherReport module that owns the shape flowing from API to display.
openmeteo.js re-exports it for backward compatibility. The OWM-compatibility
shape is now an implementation detail of the adapter, not a contract callers
memorize. Removed unused isoToUnix/nearestHourlyIndex from openmeteo.js
(they live in weatherReport.js now).

All 629 tests pass. Lint clean.
